+++ src/patch/dll_inputdisplay.zig
@@ -7,8 +7,6 @@ const GlobalFn = @import("appinfo.zig").GLOBAL_FUNCTION;
 const COMPATIBILITY_VERSION = @import("appinfo.zig").COMPATIBILITY_VERSION;
 const VERSION_STR = @import("appinfo.zig").VERSION_STR;
 
-const debug = @import("util/base/base_debug.zig");
-
 const nt = @import("util/normalized_transform.zig");
 const msg = @import("util/message.zig");
 const mem = @import("util/memory.zig");
@@ -23,8 +21,8 @@ const SettingHandle = @import("core/ASettings.zig").Handle;
 const SettingValue = @import("core/ASettings.zig").ASettingSent.Value;
 const Setting = @import("core/ASettings.zig").ASettingSent;
 
-// TODO: passthrough to annodue's panic via global function vtable; same for logging
-pub const panic = debug.annodue_panic;
+const debug_panic = @import("util/debug/debug_panic.zig");
+pub const panic = debug_panic.PanicFromContext("plugin_inputdisplay", "annodue/plugin/plugin_inputdisplay.pdb");
 
 // FEATURES
 // - Visualize inputs during race