+++ src/patch/dll_test.zig
@@ -4,8 +4,6 @@ const GlobalFn = @import("appinfo.zig").GLOBAL_FUNCTION;
 const COMPATIBILITY_VERSION = @import("appinfo.zig").COMPATIBILITY_VERSION;
 const VERSION_STR = @import("appinfo.zig").VERSION_STR;
 
-const debug = @import("util/base/base_debug.zig");
-
 const msg = @import("util/message.zig");
 
 const r = @import("racer");
@@ -15,8 +13,8 @@ const SettingHandle = @import("core/ASettings.zig").Handle;
 const SettingValue = @import("core/ASettings.zig").ASettingSent.Value;
 const Setting = @import("core/ASettings.zig").ASettingSent;
 
-// TODO: passthrough to annodue's panic via global function vtable; same for logging
-pub const panic = debug.annodue_panic;
+const debug_panic = @import("util/debug/debug_panic.zig");
+pub const panic = debug_panic.PanicFromContext("plugin_test", "annodue/plugin/plugin_test.pdb");
 
 // FEATURES
 // -
@@ -51,5 +49,8 @@ export fn OnDeinit(_: *GlobalFn) callconv(.C) void {}
 // HOOKS
 
 export fn EarlyEngineUpdateA(_: *GlobalFn) callconv(.C) void {
+    //if (gf.InputGetKb(.J, .JustOn)) std.debug.assert(false); // does nothing in ReleaseFast, ReleaseSmall
+    //if (gf.InputGetKb(.F, .JustOn)) @panic("panic test");
+
     //_ = gf.GDrawText(.Default, rt.hMakeText(0, 0, "GDrawText Test", .{}, null, null) catch null);
 }