+++ src/patch/main.zig
@@ -5,11 +5,10 @@ const AHook = @import("core/AHook.zig");
 const AMemory = @import("core/AMemory.zig");
 const ASettings = @import("core/ASettings.zig");
 
-const msg = @import("util/message.zig");
-const dbg = @import("util/base/base_debug.zig");
 const MiB = @import("util/base/base_memory.zig").MiB;
 
-pub const panic = dbg.annodue_panic;
+const debug_panic = @import("util/debug/debug_panic.zig");
+pub const panic = debug_panic.PanicFromContext("annodue", "annodue/annodue.pdb");
 
 export fn Init() void {
     // TODO: maybe this should be re-characterized to reflect that it's just